Restructured filesystem to modularize components
