+++ src/app/modules/checkout/checkout.style.css
@@ -1,0 +1,100 @@
+.checkout {
+    text-align: center;
+    color: rgb(255, 255, 255);
+  }
+  
+  .order-header {
+    text-align: center;
+    
+  }
+  
+  .box-outer {
+    height: 550px;
+    width: 500px;
+    border: 1px solid #5a5752;
+    background-color: #202020;
+    /* border: 1px solid rgb(2, 2, 2); */
+    margin: auto;
+  }
+  
+  .item-container {
+    color: rgb(255, 255, 255);
+    display: flex;
+    flex-direction: wrap;
+    justify-content: space-around;
+    height: 50px;
+    width: 400px;
+    border: 1px solid #5a5752;
+    margin: 20px;
+    padding: 20px;
+  }
+  
+  .item {
+    color: rgb(255, 255, 255);
+    text-align: center;
+    vertical-align: middle;
+    line-height: 50px;
+    width: 300px;
+    height: 50px;
+    border: 1px solid #5a5752;
+    border-radius: 18px;
+  }
+  .cost {
+    color: rgb(255, 255, 255);
+    text-align: center;
+    vertical-align: middle;
+    line-height: 50px;
+    width: 100px;
+    height: 50px;
+    border: 1px solid #5a5752;
+    margin-left: 20px;
+  }
+  
+  .additional {
+    color: rgb(255, 255, 255);
+    width: 400px;
+    height: 50px;
+    border: 1px solid #5a5752;
+    text-align: center;
+    vertical-align: middle;
+    line-height: 50px;
+    margin: 20px;
+    padding: 0 20px 0;
+  }
+  
+  .form-credit {
+    color: rgb(255, 255, 255);
+    text-align: start;
+  }
+  
+  .credit-card {
+    color: rgb(255, 255, 255);
+    width: 400px;
+    height: 250px;
+    border: 1px solid #5a5752;
+    margin: 20px;
+    padding: 0 20px 0;
+  }
+  
+  .credit-text {
+    margin: 8px;
+  }
+  
+  .button {
+    height: 40px;
+    width: auto;
+    padding-left: 10px;
+    padding-right: 10px;
+    border-radius: 10px;
+    background-color: #daab45;
+    border: 2px solid #daab45;
+    /* color: rgb(14, 13, 13); */
+    box-shadow: rgba(8, 8, 8, 0.24) 8px 8px 8px 8px;
+    text-align: center;
+    font-size: 16px;
+    font-weight: bold;
+    margin: auto 20px;
+    cursor: pointer;
+    justify-content: center;
+  }
+
