--- src/index.css
@@ -1,41 +1,0 @@
-body {
-  height: 100vh;
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-ul {
-  list-style: none;
-}
-
-a {
-  cursor: pointer;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
-
-#root {
-  height: inherit;
-}
-
-.App {
-  display: flex;
-  flex-flow: column;
-  height: 100%;
-  background: linear-gradient(#1f1f1e,#BF953F, #daab45,#BF953F, #1f1f1e 100%)
-}
-
-.App .footer {
-  flex: 0 1 40px;
-}
-
-.App .routes {
-  flex: 1 1 auto;
-}
