+++ src/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+import './css/base/_base-elements.css';
+import App from './app/App';
 import { BrowserRouter as Router } from 'react-router-dom'
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from './app/reportWebVitals';
 
 ReactDOM.render(
   <Router>
